requirements.txt
flask
pydantic
python-dotenv
rich
Flask-Caching
Flask-Migrate
Flask-SQLAlchemy
asgiref
uvicorn
google-genai
PyJWT
email_validator
Flask-WTF
Flask-Login
